requirements.txt
msgpack_python==0.5.6
pycryptodome==3.20.0
python-dotenv==1.0.1
